- Windows monitor

+++ HomeControl/WindowsMonitor/Item.cs
@@ -9,8 +9,8 @@ namespace WindowsMonitor
     public class Item
     {
         public string Name;
-        public int Value;
-        public Item(string name, int value)
+        public double Value;
+        public Item(string name, double value)
         {
             Name = name; Value = value;
         }
